chore: refactor filename and extension
